Refactor packages to separate node source windows
